refactor(vite-plugin-angular): extract template-class-binding-guard into separate plugin file
Move the template-class-binding-guard plugin, its template analysis
functions (findStaticClassAndBoundClassConflicts, findBoundClassAndNgClassConflicts,
findOpeningTagSnippets, throwTemplateClassBindingConflict), active-graph
metadata tracking, style owner registration, and related types into
template-class-binding-guard-plugin.ts.

The plugin factory receives shared state via a context object so it can
operate on the same maps used by the HMR pipeline.
