+++ packages/vite-plugin-angular/src/lib/angular-vite-plugin.ts
@@ -57,6 +57,14 @@ import type {
 
 import { compilationAPIPlugin } from './compilation-api/index.js';
 import { fastCompilePlugin } from './fast-compile-plugin.js';
+import {
+  templateClassBindingGuardPlugin,
+  removeActiveGraphMetadata,
+  removeStyleOwnerMetadata,
+  type ActiveGraphComponentRecord,
+  type StyleOwnerRecord,
+  type TemplateClassBindingGuardContext,
+} from './template-class-binding-guard-plugin.js';
 import { angularVitestPlugins } from './angular-vitest-plugin.js';
 import {
   createJitResourceTransformer,
@@ -135,6 +143,10 @@ export {
   refreshStylesheetRegistryForFile,
   describeStylesheetContent,
 } from './utils/compilation-shared.js';
+export {
+  findStaticClassAndBoundClassConflicts,
+  findBoundClassAndNgClassConflicts,
+} from './template-class-binding-guard-plugin.js';
 import {
   DiagnosticModes,
   injectViteIgnoreForHmrMetadata,
@@ -674,121 +686,6 @@ export function angular(options?: PluginOptions): Plugin[] {
     }
   }
 
-  function isLikelyPageOnlyComponent(id: string): boolean {
-    return (
-      id.includes('/pages/') ||
-      /\.page\.[cm]?[jt]sx?$/i.test(id) ||
-      /\([^/]+\)\.page\.[cm]?[jt]sx?$/i.test(id)
-    );
-  }
-
-  function removeActiveGraphMetadata(file: string) {
-    const previous = activeGraphComponentMetadata.get(file);
-    if (!previous) {
-      return;
-    }
-
-    for (const record of previous) {
-      const location = `${record.file}#${record.className}`;
-      if (record.selector) {
-        const selectorSet = selectorOwners.get(record.selector);
-        selectorSet?.delete(location);
-        if (selectorSet?.size === 0) {
-          selectorOwners.delete(record.selector);
-        }
-      }
-
-      const classNameSet = classNameOwners.get(record.className);
-      classNameSet?.delete(location);
-      if (classNameSet?.size === 0) {
-        classNameOwners.delete(record.className);
-      }
-    }
-
-    activeGraphComponentMetadata.delete(file);
-  }
-
-  function registerActiveGraphMetadata(
-    file: string,
-    records: ActiveGraphComponentRecord[],
-  ) {
-    removeActiveGraphMetadata(file);
-
-    if (records.length === 0) {
-      return;
-    }
-
-    activeGraphComponentMetadata.set(file, records);
-
-    for (const record of records) {
-      const location = `${record.file}#${record.className}`;
-
-      if (record.selector) {
-        let selectorSet = selectorOwners.get(record.selector);
-        if (!selectorSet) {
-          selectorSet = new Set<string>();
-          selectorOwners.set(record.selector, selectorSet);
-        }
-        selectorSet.add(location);
-      }
-
-      let classNameSet = classNameOwners.get(record.className);
-      if (!classNameSet) {
-        classNameSet = new Set<string>();
-        classNameOwners.set(record.className, classNameSet);
-      }
-      classNameSet.add(location);
-    }
-  }
-
-  function removeStyleOwnerMetadata(file: string) {
-    const previous = transformedStyleOwnerMetadata.get(file);
-    if (!previous) {
-      return;
-    }
-
-    for (const record of previous) {
-      const owners = styleSourceOwners.get(record.sourcePath);
-      owners?.delete(record.ownerFile);
-      if (owners?.size === 0) {
-        styleSourceOwners.delete(record.sourcePath);
-      }
-    }
-
-    transformedStyleOwnerMetadata.delete(file);
-  }
-
-  function registerStyleOwnerMetadata(file: string, styleUrls: string[]) {
-    removeStyleOwnerMetadata(file);
-
-    const records = styleUrls
-      .map((urlSet) => {
-        const [, absoluteFileUrl] = urlSet.split('|');
-        return absoluteFileUrl
-          ? {
-              ownerFile: file,
-              sourcePath: normalizePath(absoluteFileUrl),
-            }
-          : undefined;
-      })
-      .filter((record): record is StyleOwnerRecord => !!record);
-
-    if (records.length === 0) {
-      return;
-    }
-
-    transformedStyleOwnerMetadata.set(file, records);
-
-    for (const record of records) {
-      let owners = styleSourceOwners.get(record.sourcePath);
-      if (!owners) {
-        owners = new Set<string>();
-        styleSourceOwners.set(record.sourcePath, owners);
-      }
-      owners.add(record.ownerFile);
-    }
-  }
-
   let stylesheetRegistry: AnalogStylesheetRegistry | undefined;
   const sourceFileCache: SourceFileCacheType = new SourceFileCache();
   const isTest = process.env['NODE_ENV'] === 'test' || !!process.env['VITEST'];
@@ -801,6 +698,14 @@ export function angular(options?: PluginOptions): Plugin[] {
   let viteServer: ViteDevServer | undefined;
 
   const styleUrlsResolver = new StyleUrlsResolver();
+  const guardContext: TemplateClassBindingGuardContext = {
+    styleUrlsResolver,
+    activeGraphComponentMetadata,
+    selectorOwners,
+    classNameOwners,
+    transformedStyleOwnerMetadata,
+    styleSourceOwners,
+  };
   const templateUrlsResolver = new TemplateUrlsResolver();
   let outputFile: ((file: string) => void) | undefined;
   const outputFiles = new Map<string, EmitFileResult>();
@@ -1001,8 +906,10 @@ export function angular(options?: PluginOptions): Plugin[] {
         server.watcher.on('add', invalidateCompilationOnFsChange);
         server.watcher.on('unlink', (file) => {
           evictDeletedFileMetadata(file, {
-            removeActiveGraphMetadata,
-            removeStyleOwnerMetadata,
+            removeActiveGraphMetadata: (f) =>
+              removeActiveGraphMetadata(guardContext, f),
+            removeStyleOwnerMetadata: (f) =>
+              removeStyleOwnerMetadata(guardContext, f),
             classNamesMap: classNames as Map<string, string>,
             fileTransformMap,
           });
@@ -1949,142 +1856,7 @@ export function angular(options?: PluginOptions): Plugin[] {
 
   return [
     replaceFiles(pluginOptions.fileReplacements, pluginOptions.workspaceRoot),
-    {
-      name: '@analogjs/vite-plugin-angular:template-class-binding-guard',
-      enforce: 'pre',
-      transform(code: string, id: string) {
-        if (id.includes('node_modules')) {
-          return;
-        }
-
-        const cleanId = id.split('?')[0];
-
-        if (/\.(html|htm)$/i.test(cleanId)) {
-          const staticClassIssue =
-            findStaticClassAndBoundClassConflicts(code)[0];
-          if (staticClassIssue) {
-            throwTemplateClassBindingConflict(cleanId, staticClassIssue);
-          }
-
-          const mixedClassIssue = findBoundClassAndNgClassConflicts(code)[0];
-          if (mixedClassIssue) {
-            this.warn(
-              [
-                '[Analog Angular] Conflicting class composition.',
-                `File: ${cleanId}:${mixedClassIssue.line}:${mixedClassIssue.column}`,
-                'This element mixes `[class]` and `[ngClass]`.',
-                'Prefer a single class-binding strategy so class merging stays predictable.',
-                'Use one `[ngClass]` expression or explicit `[class.foo]` bindings.',
-                `Snippet: ${mixedClassIssue.snippet}`,
-              ].join('\n'),
-            );
-          }
-          return;
-        }
-
-        if (TS_EXT_REGEX.test(cleanId)) {
-          const rawStyleUrls = styleUrlsResolver.resolve(code, cleanId);
-          registerStyleOwnerMetadata(cleanId, rawStyleUrls);
-          debugHmrV('component stylesheet owner metadata registered', {
-            file: cleanId,
-            styleUrlCount: rawStyleUrls.length,
-            styleUrls: rawStyleUrls,
-            ownerSources: [
-              ...(transformedStyleOwnerMetadata
-                .get(cleanId)
-                ?.map((record) => record.sourcePath) ?? []),
-            ],
-          });
-
-          // Parse raw component decorators before Angular compilation strips
-          // them. This lets Analog fail fast on template/class-footguns and
-          // keep a lightweight active-graph index for duplicate selector/class
-          // diagnostics without requiring a full compiler pass first.
-          const components = getAngularComponentMetadata(code);
-
-          const inlineTemplateIssue = components.flatMap((component) =>
-            component.inlineTemplates.flatMap((template) =>
-              findStaticClassAndBoundClassConflicts(template),
-            ),
-          )[0];
-
-          if (inlineTemplateIssue) {
-            throwTemplateClassBindingConflict(cleanId, inlineTemplateIssue);
-          }
-
-          const mixedInlineClassIssue = components.flatMap((component) =>
-            component.inlineTemplates.flatMap((template) =>
-              findBoundClassAndNgClassConflicts(template),
-            ),
-          )[0];
-
-          if (mixedInlineClassIssue) {
-            this.warn(
-              [
-                '[Analog Angular] Conflicting class composition.',
-                `File: ${cleanId}:${mixedInlineClassIssue.line}:${mixedInlineClassIssue.column}`,
-                'This element mixes `[class]` and `[ngClass]`.',
-                'Prefer a single class-binding strategy so class merging stays predictable.',
-                'Use one `[ngClass]` expression or explicit `[class.foo]` bindings.',
-                `Snippet: ${mixedInlineClassIssue.snippet}`,
-              ].join('\n'),
-            );
-          }
-
-          const activeGraphRecords = components.map((component) => ({
-            file: cleanId,
-            className: component.className,
-            selector: component.selector,
-          }));
-
-          registerActiveGraphMetadata(cleanId, activeGraphRecords);
-
-          for (const component of components) {
-            if (!component.selector && !isLikelyPageOnlyComponent(cleanId)) {
-              throw new Error(
-                [
-                  '[Analog Angular] Selectorless component detected.',
-                  `File: ${cleanId}`,
-                  `Component: ${component.className}`,
-                  'This component has no `selector`, so Angular will render it as `ng-component`.',
-                  'That increases the chance of component ID collisions and makes diagnostics harder to interpret.',
-                  'Add an explicit selector for reusable components.',
-                  'Selectorless components are only supported for page and route-only files.',
-                ].join('\n'),
-              );
-            }
-
-            if (component.selector) {
-              const selectorEntries = selectorOwners.get(component.selector);
-              if (selectorEntries && selectorEntries.size > 1) {
-                throw new Error(
-                  [
-                    '[Analog Angular] Duplicate component selector detected.',
-                    `Selector: ${component.selector}`,
-                    'Multiple components in the active application graph use the same selector.',
-                    'Selectors must be unique within the active graph to avoid ambiguous rendering and confusing diagnostics.',
-                    `Locations:\n${formatActiveGraphLocations(selectorEntries)}`,
-                  ].join('\n'),
-                );
-              }
-            }
-
-            const classNameEntries = classNameOwners.get(component.className);
-            if (classNameEntries && classNameEntries.size > 1) {
-              this.warn(
-                [
-                  '[Analog Angular] Duplicate component class name detected.',
-                  `Class name: ${component.className}`,
-                  'Two or more Angular components in the active graph share the same exported class name.',
-                  'Rename one of them to keep HMR, stack traces, and compiler diagnostics unambiguous.',
-                  `Locations:\n${formatActiveGraphLocations(classNameEntries)}`,
-                ].join('\n'),
-              );
-            }
-          }
-        }
-      },
-    } satisfies Plugin,
+    templateClassBindingGuardPlugin(guardContext),
     // Tailwind CSS v4 @reference injection for direct-file-loaded CSS.
     // Catches CSS files loaded from disk (not virtual modules) that need
     // @reference before @tailwindcss/vite processes them.
@@ -3249,158 +3021,10 @@ function resolveComponentClassNamesForStyleOwner(
     .map((component) => component.className);
 }
 
-interface TemplateClassBindingIssue {
-  line: number;
-  column: number;
-  snippet: string;
-}
-
-interface ActiveGraphComponentRecord {
-  file: string;
-  className: string;
-  selector?: string;
-}
-
-interface StyleOwnerRecord {
-  sourcePath: string;
-  ownerFile: string;
-}
-
 type ComponentStylesheetHmrOutcome =
   | 'css-update'
   | 'owner-component-update'
   | 'full-reload';
-
-export function findStaticClassAndBoundClassConflicts(
-  template: string,
-): TemplateClassBindingIssue[] {
-  const issues: TemplateClassBindingIssue[] = [];
-
-  for (const { index, snippet } of findOpeningTagSnippets(template)) {
-    if (!snippet.includes('[class]')) {
-      continue;
-    }
-
-    const hasStaticClass = /\sclass\s*=\s*(['"])(?:(?!\1)[\s\S])*\1/.test(
-      snippet,
-    );
-    const hasBoundClass = /\s\[class\]\s*=\s*(['"])(?:(?!\1)[\s\S])*\1/.test(
-      snippet,
-    );
-
-    if (hasStaticClass && hasBoundClass) {
-      const prefix = template.slice(0, index);
-      const line = prefix.split('\n').length;
-      const lastNewline = prefix.lastIndexOf('\n');
-      const column = index - lastNewline;
-      issues.push({
-        line,
-        column,
-        snippet: snippet.replace(/\s+/g, ' ').trim(),
-      });
-    }
-  }
-
-  return issues;
-}
-
-function throwTemplateClassBindingConflict(
-  id: string,
-  issue: TemplateClassBindingIssue,
-): never {
-  throw new Error(
-    [
-      '[Analog Angular] Invalid template class binding.',
-      `File: ${id}:${issue.line}:${issue.column}`,
-      'The same element uses both a static `class="..."` attribute and a whole-element `[class]="..."` binding.',
-      'That pattern can replace or conflict with static Tailwind classes, which makes styles appear to stop applying.',
-      'Use `[ngClass]` or explicit `[class.foo]` bindings instead of `[class]` when the element also has static classes.',
-      `Snippet: ${issue.snippet}`,
-    ].join('\n'),
-  );
-}
-
-export function findBoundClassAndNgClassConflicts(
-  template: string,
-): TemplateClassBindingIssue[] {
-  const issues: TemplateClassBindingIssue[] = [];
-  const hasWholeElementClassBinding = /\[class\]\s*=/.test(template);
-
-  if (!hasWholeElementClassBinding || !template.includes('[ngClass]')) {
-    return issues;
-  }
-
-  for (const { index, snippet } of findOpeningTagSnippets(template)) {
-    if (!/\[class\]\s*=/.test(snippet) || !snippet.includes('[ngClass]')) {
-      continue;
-    }
-
-    const prefix = template.slice(0, index);
-    const line = prefix.split('\n').length;
-    const lastNewline = prefix.lastIndexOf('\n');
-    const column = index - lastNewline;
-    issues.push({
-      line,
-      column,
-      snippet: snippet.replace(/\s+/g, ' ').trim(),
-    });
-  }
-
-  return issues;
-}
-
-function findOpeningTagSnippets(
-  template: string,
-): Array<{ index: number; snippet: string }> {
-  const matches: Array<{ index: number; snippet: string }> = [];
-
-  for (let index = 0; index < template.length; index++) {
-    if (template[index] !== '<') {
-      continue;
-    }
-
-    const tagStart = template[index + 1];
-    if (!tagStart || !/[a-zA-Z]/.test(tagStart)) {
-      continue;
-    }
-
-    let quote: '"' | "'" | null = null;
-
-    for (let end = index + 1; end < template.length; end++) {
-      const char = template[end];
-
-      if (quote) {
-        if (char === quote) {
-          quote = null;
-        }
-        continue;
-      }
-
-      if (char === '"' || char === "'") {
-        quote = char;
-        continue;
-      }
-
-      if (char === '>') {
-        matches.push({
-          index,
-          snippet: template.slice(index, end + 1),
-        });
-        index = end;
-        break;
-      }
-    }
-  }
-
-  return matches;
-}
-
-function formatActiveGraphLocations(entries: Iterable<string>): string {
-  return [...entries]
-    .sort()
-    .map((entry) => `- ${entry}`)
-    .join('\n');
-}
 
 function logComponentStylesheetHmrOutcome(details: {
   file: string;
